Thêm phần mua ngay sản phẩm cho customer và cập nhập thêm lịch sử mua hàng

+++ Sales Management/Models/Product.cs
@@ -24,7 +24,7 @@ public partial class Product
     [Range(0.01, double.MaxValue, ErrorMessage = "Giá bán phải lớn hơn 0")]
     public decimal SellingPrice { get; set; }
 
-    public int? CoinPrice { get; set; }
+    public decimal? CoinPrice { get; set; }
 
     public decimal? Vatrate { get; set; }
